pyproject.toml
[project]
name = "hackatl2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.193.2",
    "crewai-tools>=0.76.0",
    "fastapi>=0.119.0",
    "motor>=3.7.1",
    "openai>=1.109.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
